DELETE: Status and StatusCheck from common
