+++ src/lib/kotlin/slatekit-jobs/src/main/kotlin/slatekit/jobs/support/Coordinator.kt
@@ -1,6 +1,6 @@
 package slatekit.jobs.support
 
-import slatekit.common.Status
+import slatekit.actors.Status
 import slatekit.actors.Action
 import slatekit.jobs.Job
 import slatekit.jobs.workers.Work
