+++ src/lib/kotlin/slatekit-jobs/src/main/kotlin/slatekit/jobs/workers/Work.kt
@@ -1,7 +1,7 @@
 package slatekit.jobs.workers
 
 import slatekit.common.Identity
-import slatekit.common.Status
+import slatekit.actors.Status
 import slatekit.actors.Action
 import slatekit.jobs.Job
 import slatekit.jobs.Task
