+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/common/StatusTests.kt
@@ -3,7 +3,7 @@ package test.common
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert
 import org.junit.Test
-import slatekit.common.Status
+import slatekit.actors.Status
 import test.setup.MyWorker
