+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/jobs/Job_Manage_Tests.kt
@@ -3,7 +3,7 @@ package test.jobs
 import org.junit.Assert
 import org.junit.Test
 import slatekit.common.Identity
-import slatekit.common.Status
+import slatekit.actors.Status
 import test.jobs.samples.TestWorker
 import test.jobs.support.JobTestSupport
