+++ src/lib/kotlin/slatekit/src/main/resources/templates/slatekit/job/files/src/Jobs.txt
@@ -2,7 +2,7 @@ package ${app.package}
 
 import kotlinx.coroutines.runBlocking
 import slatekit.common.Identity
-import slatekit.common.Status
+import slatekit.actors.Status
 import slatekit.jobs.Task
 import slatekit.jobs.workers.WorkResult
 import slatekit.jobs.workers.Worker
